[refact] ♻ 스케쥴링을 위한 코드 정비
- QuartzConfiguration.java는 실제로 자기 자신만의 스케쥴을 가지고 있는거지, 배치잡에 영향을 주진 않음. 일단 주석처리
- MovieBatchConfig로 이름 변경
- 필요없는 로거 삭제
