+++ src/main/java/com/ventulus95/springbatch/scheduler/ScheduledJob.java
@@ -2,12 +2,11 @@ package com.ventulus95.springbatch.scheduler;
 
 
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.quartz.JobDataMap;
 import org.quartz.JobExecutionContext;
 import org.quartz.JobExecutionException;
 import org.quartz.SchedulerException;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 import org.springframework.batch.core.JobParameters;
 import org.springframework.batch.core.JobParametersBuilder;
 import org.springframework.batch.core.JobParametersInvalidException;
@@ -19,30 +18,29 @@ import org.springframework.batch.core.repository.JobInstanceAlreadyCompleteExcep
 import org.springframework.batch.core.repository.JobRestartException;
 import org.springframework.scheduling.quartz.QuartzJobBean;
 
+@Slf4j
 @RequiredArgsConstructor
 public class ScheduledJob extends QuartzJobBean {
 
-    private final Logger logger = LoggerFactory.getLogger(ScheduledJob.class);
     private final JobLocator jobLocator;
     private final JobLauncher jobLauncher;
 
 
     @Override //excuteInternal는 이벤트 발생시마다 한번씩 호출되니까 스케쥴할 무엇인가만 여기에 얹어주면 된다.
     protected void executeInternal(JobExecutionContext context) throws JobExecutionException {
-        logger.info("excueteInternal!!! START!!");
+        log.info("excueteInternal!!! START!!");
         JobDataMap map = context.getMergedJobDataMap();
         JobParameters parameters = null;
         try {
             parameters = new JobParametersBuilder()
                     .addString("InstanceId", context.getScheduler().getSchedulerInstanceId())
                     .addLong("timestamp", System.currentTimeMillis())
-                    .toJobParameters();
+                    .toJobParameters(); //파라미터를 instanceId, timestamp으로 설정
 
-            logger.info("job은 뭐고... :{}, ",(String) map.get("job"));
+            log.info("job은 뭐고... :{}, ",(String) map.get("job"));
             jobLauncher.run(jobLocator.getJob((String) map.get("job")), parameters);
-            logger.info("[{}] 배치 잡 완료", (String) map.get("job"));
         } catch (JobExecutionAlreadyRunningException | JobInstanceAlreadyCompleteException | JobParametersInvalidException| NoSuchJobException| JobRestartException | SchedulerException  e) {
-            e.printStackTrace();
+            log.error(e.toString());
             throw new JobExecutionException();
         }
     }
